refactor: AddCardActivity로 액티비티명 변경
